@@ -11,46 +11,46 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.1.4",
-    "@angular/cdk": "^7.2.0",
-    "@angular/common": "^7.1.4",
-    "@angular/compiler": "^7.1.4",
-    "@angular/core": "^7.1.4",
-    "@angular/flex-layout": "^7.0.0-beta.22",
-    "@angular/forms": "^7.1.4",
+    "@angular/animations": "^8.2.0",
+    "@angular/cdk": "^8.1.2",
+    "@angular/common": "^8.2.0",
+    "@angular/compiler": "^8.2.0",
+    "@angular/core": "^8.2.0",
+    "@angular/flex-layout": "^8.0.0-beta.26",
+    "@angular/forms": "^8.2.0",
     "@angular/http": "^7.1.4",
-    "@angular/material": "~7.2.0",
-    "@angular/platform-browser": "^7.1.4",
-    "@angular/platform-browser-dynamic": "^7.1.4",
-    "@angular/router": "^7.1.4",
-    "core-js": "^2.6.1",
+    "@angular/material": "~8.1.2",
+    "@angular/platform-browser": "^8.2.0",
+    "@angular/platform-browser-dynamic": "^8.2.0",
+    "@angular/router": "^8.2.0",
+    "core-js": "^3.1.4",
     "rxjs": "^6.3.3",
     "tslib": "^1.9.0",
-    "zone.js": "^0.8.26",
+    "zone.js": "^0.10.0",
     "socket.io": "^2.2.0",
     "@types/socket.io": "latest",
     "@types/express": "latest",
     "express": "latest"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.11.0",
-    "@angular/cli": "~7.1.4",
-    "@angular/compiler-cli": "^7.1.4",
-    "@angular/language-service": "^7.1.4",
-    "@types/jasmine": "^2.8.14",
+    "@angular-devkit/build-angular": "~0.802.0",
+    "@angular/cli": "~8.2.0",
+    "@angular/compiler-cli": "^8.2.0",
+    "@angular/language-service": "^8.2.0",
+    "@types/jasmine": "^3.3.16",
     "@types/jasminewd2": "^2.0.6",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.2.1",
-    "jasmine-core": "~2.99.1",
+    "@types/node": "~12.6.8",
+    "codelyzer": "~5.1.0",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~1.7.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.0",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.3.0",
-    "ts-node": "~5.0.1",
-    "tslint": "~5.9.1",
-    "typescript": "^3.1.6"
+    "karma": "~4.2.0",
+    "karma-chrome-launcher": "~3.0.0",
+    "karma-coverage-istanbul-reporter": "~2.1.0",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
+    "protractor": "~5.4.2",
+    "ts-node": "~8.3.0",
+    "tslint": "~5.18.0",
+    "typescript": "^3.4.0"
   }
 }
