@@ -5,7 +5,7 @@ import { MatSidenav } from '@angular/material';
 //import { ContentScrollListenerService } from '../services/contentscrolllistener.service';
 
 //ContentScrollStuff
-import { DOCUMENT } from '@angular/platform-browser';
+import { DOCUMENT } from '@angular/common';
 
 @Component({
   selector: 'app-header',
@@ -18,6 +18,7 @@ export class HeaderComponent implements OnInit, AfterViewInit {
 
   public contentScrollEvent;
   public currentScrollPosition;
+  //public elem;
 
   constructor(public toggleService: ToggleService,
               public themeService : ThemeService,
@@ -25,16 +26,20 @@ export class HeaderComponent implements OnInit, AfterViewInit {
               @Inject(DOCUMENT) private document: Document) { }
 
   ngOnInit() {
+    //     this.elem = document.getElementsByClassName('parallax-bg')[0].getBoundingClientRect();
+
+    // //Constantly listen to content scrolling inside "mat-sidenav-content"
+    // this.document.querySelector('mat-sidenav-content')
+    //                              .addEventListener('scroll', this.onContentScroll.bind(this));
 
   }
 
   ngAfterViewInit() {
-
+    //this.elem = document.getElementsByClassName('parallax-bg')[0].getBoundingClientRect();
     //Possibly don't need below lines
-
-    //Constantly listen to content scrolling inside "mat-sidenav-content"
     this.document.querySelector('mat-sidenav-content')
                                  .addEventListener('scroll', this.onContentScroll.bind(this));
+    console.log("[HEADER] ngAfterViewInit");
   }
 
   toggleActive:boolean = false;
@@ -52,23 +57,24 @@ export class HeaderComponent implements OnInit, AfterViewInit {
 
   @Input() public index: number;
   public isElevated: boolean = false;
-
   public switchedOn :boolean = true;
   public elevationValue = 0
 
   onContentScroll(event) {
-    //console.log("[header scroll]");
-    //console.log("[header.component : onContentScroll] ");
-    //console.dir(event);
     let scrolled = window.pageYOffset;
     let rect = this.elementRef.nativeElement.getBoundingClientRect().top;
     let elementOffsetTop = this.elementRef.nativeElement.offsetTop;
 
     let elem = document.getElementsByClassName('parallax-bg')[0].getBoundingClientRect();
-    //console.dir(elem);
+
+    // console.log("[header] elem.top " + elem.top);
+    // console.dir(elem);
 
     if( elem.top < 64 ) 
     {
+    //       console.log("[header scroll]");
+    // console.log("[header.component : onContentScroll] ");
+    // console.dir(event);
       if(this.isElevated) {
         this.isElevated = false;
       }
