@@ -1,134 +1,50 @@
 import { isPlatformBrowser } from '@angular/common';
-import { AfterViewInit, Directive, ElementRef, EventEmitter, Inject, Input, NgZone, OnDestroy, Output, PLATFORM_ID } from '@angular/core';
-import { fromEvent, Subscription } from 'rxjs';
-import { debounceTime } from 'rxjs/operators';;
+import { AfterViewInit, Directive, ElementRef, HostBinding, Input } from '@angular/core';
 
+
+/**
+  Defer Load Directive
+  Lazy Loading images:
+    Instead of loading these resources as soon as the page loads, which is what normally happens, 
+    the loading of these resources is put off till the moment the user actually needs to view them.
+
+  https://imagekit.io/blog/lazy-loading-images-complete-guide/
+  https://medium.com/@realTomaszKula/lazy-load-images-in-30-lines-of-code-3fe801223ffa
+
+*/
 @Directive({
-  selector: '[deferLoad]'
+  selector: '[lazy-load]'
 })
-export class LazyloadDirective implements AfterViewInit, OnDestroy {
-  @Input() public preRender: boolean = true;
-    @Output() public deferLoad: EventEmitter<any> = new EventEmitter();
+export class LazyloadDirective implements AfterViewInit {
+  @HostBinding('attr.src') srcAttr = null;
+  @Input() src: string;
 
-    private _intersectionObserver?: IntersectionObserver;
-    private _scrollSubscription?: Subscription;
+  constructor(private el: ElementRef) {}
 
-    constructor (
-        private _element: ElementRef,
-        private _zone: NgZone,
-        @Inject(PLATFORM_ID) private platformId: Object
-    ) { }
+  ngAfterViewInit() {
+    console.log("[LAZYLOAD] canLazyLoad: " + this.canLazyLoad());
+    this.canLazyLoad() ? this.lazyLoadImage() : this.loadImage();
+  }
 
-    public ngAfterViewInit () {
-        if (isPlatformBrowser(this.platformId)) {
-            if (this.hasCompatibleBrowser()) {
-                this.registerIntersectionObserver();
-                if (this._intersectionObserver && this._element.nativeElement) {
-                    this._intersectionObserver.observe(<Element>(this._element.nativeElement));
-                }
-            } else {
-                this.addScrollListeners();
-            }
-        } else {
-            if (this.preRender) {
-                this.load();
-            }
+  private canLazyLoad() {
+    //console.log("[LAZYLOAD] canLazyLoad: " + (window && 'IntersectionObserver' in window));
+    return window && 'IntersectionObserver' in window;
+  }
+
+  private lazyLoadImage() {
+    const obs = new IntersectionObserver(entries => {
+      entries.forEach(({ isIntersecting }) => {
+        if (isIntersecting) {
+          console.log("[IMAGE LOADED]");
+          this.loadImage();
+          obs.unobserve(this.el.nativeElement);
         }
-    }
+      });
+    });
+    obs.observe(this.el.nativeElement);
+  }
 
-    public hasCompatibleBrowser (): boolean {
-        const hasIntersectionObserver = 'IntersectionObserver' in window;
-        const userAgent = window.navigator.userAgent;
-        const matches = userAgent.match(/Edge\/(\d*)\./i);
-
-        const isEdge = !!matches && matches.length > 1;
-        const isEdgeVersion16OrBetter = isEdge && (!!matches && parseInt(matches[1], 10) > 15);
-
-        return hasIntersectionObserver && (!isEdge || isEdgeVersion16OrBetter);
-    }
-
-    public ngOnDestroy () {
-        this.removeListeners();
-    }
-
-    private registerIntersectionObserver (): void {
-        if (!!this._intersectionObserver) {
-            return;
-        }
-        this._intersectionObserver = new IntersectionObserver(entries => {
-              console.log("[LazyLoad] registerIntersectionObserver");       //DEBUGGING
-            this.checkForIntersection(entries);
-        }, {});
-    }
-
-    private checkForIntersection = (entries: Array<IntersectionObserverEntry>) => {
-        entries.forEach((entry: IntersectionObserverEntry) => {
-            console.log("[LazyLoad] checkIfIntersection");  //DEBUGGING
-
-            if (this.checkIfIntersecting(entry)) {
-                this.load();
-                if (this._intersectionObserver && this._element.nativeElement) {
-                    this._intersectionObserver.unobserve(<Element>(this._element.nativeElement));
-                }
-            }
-        });
-    }
-
-    private checkIfIntersecting (entry: IntersectionObserverEntry) {
-
-        console.log("[LazyLoad] checkIfIntersecting");  //DEBUGGING
-
-        // For Samsung native browser, IO has been partially implemented where by the
-        // callback fires, but entry object is empty. We will check manually.
-        if (entry && Object.keys(entry).length) {
-            return (<any>entry).isIntersecting && entry.target === this._element.nativeElement;
-        }
-        return this.isVisible();
-    }
-
-    private load (): void {
-        console.log("[LazyLoad] load"); //DEBUGGING
-        this.removeListeners();
-        this.deferLoad.emit(true);
-    }
-
-    private addScrollListeners () {
-        if (this.isVisible()) {
-            this.load();
-            return;
-        }
-        this._zone.runOutsideAngular(() => {
-            this._scrollSubscription = fromEvent(window, 'scroll')
-                .pipe(debounceTime(50))
-                .subscribe(this.onScroll);
-        });
-    }
-
-    private removeListeners () {
-        if (this._scrollSubscription) {
-            this._scrollSubscription.unsubscribe();
-        }
-
-        if (this._intersectionObserver) {
-            this._intersectionObserver.disconnect();
-        }
-    }
-
-    private onScroll = () => {
-        if (this.isVisible()) {
-            this._zone.run(() => this.load());
-        }
-    }
-
-    private isVisible () {
-        let scrollPosition = this.getScrollPosition();
-        let elementOffset = this._element.nativeElement.offsetTop;
-        return elementOffset <= scrollPosition;
-    }
-
-    private getScrollPosition () {
-        // Getting screen size and scroll position for IE
-        return (window.scrollY || window.pageYOffset)
-            + (document.documentElement.clientHeight || document.body.clientHeight);
-    }
+  private loadImage() {
+    this.srcAttr = this.src;
+  }
 }
