@@ -1,5 +1,5 @@
 import { Injectable, Inject, Input } from '@angular/core';
-import { DOCUMENT } from '@angular/platform-browser';
+import { DOCUMENT } from '@angular/common';
 import { BehaviorSubject ,  fromEvent } from 'rxjs';
 import { map, pairwise, filter, exhaustMap } from 'rxjs/operators';
 
